refactor: NotificationService를 useNotificationMonitor 훅으로 이관 및 React 생명주기 통합 (#2)
